调整代码函数

+++ model/application_usage.go
@@ -4,7 +4,7 @@ package model
 type ApplicationUsage struct {
 	Base
 	ApplicationId  int64  `gorm:"not null;index" json:"application_id"`  // 组织应用ID
-	UserId         *uint  `gorm:"index" json:"user_id"`                  // 用户ID，可为空表示系统操作
+	UserId         *int64 `gorm:"index" json:"user_id"`                  // 用户ID，可为空表示系统操作
 	UsageType      string `gorm:"size:50;not null" json:"usage_type"`    // 使用类型：api_call, storage, feature
 	UsageAmount    int64  `gorm:"not null" json:"usage_amount"`          // 使用量
 	UsageDate      int64  `gorm:"not null;index" json:"usage_date"`      // 使用日期
