+++ repository/organization_member_repository.go
@@ -9,13 +9,13 @@ import (
 // OrganizationMemberRepository 组织成员仓库接口
 type OrganizationMemberRepository interface {
 	Create(ctx context.Context, member *model.OrganizationMember) error
-	GetByID(ctx context.Context, id uint) (*model.OrganizationMember, error)
-	GetByOrganizationAndUser(ctx context.Context, orgID, userID uint) (*model.OrganizationMember, error)
-	GetByOrganization(ctx context.Context, orgID uint, page, pageSize int) ([]model.OrganizationMember, int64, error)
-	GetByUser(ctx context.Context, userID uint) ([]model.OrganizationMember, error)
+	GetByID(ctx context.Context, id int64) (*model.OrganizationMember, error)
+	GetByOrganizationAndUser(ctx context.Context, orgID, userID int64) (*model.OrganizationMember, error)
+	GetByOrganization(ctx context.Context, orgID int64, page, pageSize int) ([]model.OrganizationMember, int64, error)
+	GetByUser(ctx context.Context, userID int64) ([]model.OrganizationMember, error)
 	Update(ctx context.Context, member *model.OrganizationMember) error
-	Delete(ctx context.Context, id uint) error
-	DeleteByOrganizationAndUser(ctx context.Context, orgID, userID uint) error
+	Delete(ctx context.Context, id int64) error
+	DeleteByOrganizationAndUser(ctx context.Context, orgID, userID int64) error
 }
 
 // organizationMemberRepository 组织成员仓库实现
@@ -32,7 +32,7 @@ func (r *organizationMemberRepository) Create(ctx context.Context, member *model
 }
 
 // GetByID 根据ID获取组织成员
-func (r *organizationMemberRepository) GetByID(ctx context.Context, id uint) (*model.OrganizationMember, error) {
+func (r *organizationMemberRepository) GetByID(ctx context.Context, id int64) (*model.OrganizationMember, error) {
 	var member model.OrganizationMember
 	err := config.DB.WithContext(ctx).First(&member, id).Error
 	if err != nil {
@@ -42,7 +42,7 @@ func (r *organizationMemberRepository) GetByID(ctx context.Context, id uint) (*m
 }
 
 // GetByOrganizationAndUser 根据组织ID和用户ID获取组织成员
-func (r *organizationMemberRepository) GetByOrganizationAndUser(ctx context.Context, orgID, userID uint) (*model.OrganizationMember, error) {
+func (r *organizationMemberRepository) GetByOrganizationAndUser(ctx context.Context, orgID, userID int64) (*model.OrganizationMember, error) {
 	var member model.OrganizationMember
 	err := config.DB.WithContext(ctx).Where("organization_id = ? AND user_id = ?", orgID, userID).First(&member).Error
 	if err != nil {
@@ -52,7 +52,7 @@ func (r *organizationMemberRepository) GetByOrganizationAndUser(ctx context.Cont
 }
 
 // GetByOrganization 根据组织ID获取成员列表
-func (r *organizationMemberRepository) GetByOrganization(ctx context.Context, orgID uint, page, pageSize int) ([]model.OrganizationMember, int64, error) {
+func (r *organizationMemberRepository) GetByOrganization(ctx context.Context, orgID int64, page, pageSize int) ([]model.OrganizationMember, int64, error) {
 	var members []model.OrganizationMember
 	var total int64
 
@@ -76,7 +76,7 @@ func (r *organizationMemberRepository) GetByOrganization(ctx context.Context, or
 }
 
 // GetByUser 根据用户ID获取所属组织列表
-func (r *organizationMemberRepository) GetByUser(ctx context.Context, userID uint) ([]model.OrganizationMember, error) {
+func (r *organizationMemberRepository) GetByUser(ctx context.Context, userID int64) ([]model.OrganizationMember, error) {
 	var members []model.OrganizationMember
 	err := config.DB.WithContext(ctx).Where("user_id = ?", userID).Find(&members).Error
 	if err != nil {
@@ -91,12 +91,12 @@ func (r *organizationMemberRepository) Update(ctx context.Context, member *model
 }
 
 // Delete 删除组织成员（软删除）
-func (r *organizationMemberRepository) Delete(ctx context.Context, id uint) error {
+func (r *organizationMemberRepository) Delete(ctx context.Context, id int64) error {
 	return config.DB.WithContext(ctx).Delete(&model.OrganizationMember{}, id).Error
 }
 
 // DeleteByOrganizationAndUser 根据组织ID和用户ID删除组织成员
-func (r *organizationMemberRepository) DeleteByOrganizationAndUser(ctx context.Context, orgID, userID uint) error {
+func (r *organizationMemberRepository) DeleteByOrganizationAndUser(ctx context.Context, orgID, userID int64) error {
 	return config.DB.WithContext(ctx).Where("organization_id = ? AND user_id = ?", orgID, userID).
 		Delete(&model.OrganizationMember{}).Error
 }
